requirements.txt
flask
mysql-python
flask-sqlalchemy
flask-script
flask-wtf
flask-migrate
py-bcrypt
python-slugify
flask-markdown
flask-uploads
Flask-Social
http://github.com/pythonforfacebook/facebook-sdk/tarball/master
python-twitter
foursquare
google-api-python-client
Flask-OAuth